Handle File arguments as filepath if value exists
This is useful for flags like `kubeconfig`, if `value` exists on the
filesystem then that file is used, otherwise the content is written to a
temporary file.

This will only work for paths without expansions or environment
variables (e.g. `$HOME`, `$(pwd)` or `~`)

if a file in the `/tmp` directory is passed it will be deleted after
running this action.

Closes #2